finished global refactoring
