+++ src/components/Modals/AddCategory/AddCategory.jsx
@@ -1,10 +1,11 @@
-import s from "./AddCategory.module.css"
+import s1 from "./AddCategory.module.css"
+import s from "../Modals.module.css"
 
 const AddCategory = (props) => {
   return (
     <div className={[s.modal, s.modal_categoryAdd].join(" ")} ref={props.modalRef}>
       <div className={s.modalcontent}>
-        <input className={[s.categoryName, s.categoryNamemodificate].join(" ")}
+        <input className={[s.categoryName, s1.categoryNamemodificate].join(" ")}
           maxlength='50'
           type="text"
           value={props.categoryName}
