+++ src/components/Modals/EditCategory/EditCategory.module.css
@@ -1,86 +1,4 @@
-.modal {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  display: block;
-  color: #fff;
-  background-color: #a5cab6;
-  z-index: 10;
-  width: 500px;
-  height: 570px;
-  border-radius: 30px;
-  overflow: auto;
-}
-
-.modal_categoryAdd {
-  width: 400px;
-  height: 360px;
-}
-
-.modalcontent {
-  margin-top: 18px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-}
-
-.categoryName {
-  width: 100%;
-  height: 30px;
-  text-align: left;
-  padding-left: 0;
-  color: #000;
-  border: 0;
-  outline: none;
-  background-color: #a5cab6;
-  border-bottom: 1px solid #000;
-  width: 300px;
-}
-
-.categoryName::placeholder {
-  color: #000;
-  font-size: 16px;
-  font-weight: normal;
-  padding: 0;
-}
-
-.categoryName:focus {
-  border-color: #007bff;
-}
-
 .categoryNamemodificate {
   text-align: center;
 }
 
-.description_color {
-  color: #000;
-  font-size: 16px;
-  margin-bottom: 20px;
-}
-
-input[type="color"] {
-  width: 50px;
-  height: 25px;
-  cursor: pointer;
-}
-
-.modalcontent input {
-  margin-bottom: 20px;
-  padding-left: 0px;
-}
-
-.closeModalCategory {
-  background-color: #CB1C1C;
-  color: #fff;
-  width: 200px;
-  height: 40px;
-  border-radius: 10px;
-  margin-bottom: 10px;
-  transition: filter 0.3s ease;
-}
-
-.closeModalCategory:hover {
-  filter: brightness(1.2);
-}
