+++ src/components/Modals/KirillLoh/KirillLoh.module.css
@@ -1,30 +1,6 @@
-
-.modal {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  display: block;
-  color: #fff;
-  background-color: #a5cab6;
-  z-index: 10;
-  width: 500px;
-  height: 570px;
-  border-radius: 30px;
-  overflow: auto;
-}
-
 .Kirillloh {
   width: 400px;
   height: 370px;
-}
-
-.modalcontent {
-  margin-top: 18px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
 }
 
 .modalcontent img {
@@ -38,46 +14,3 @@
   margin-bottom: 20px;
 }
 
-.close {
-  background-color: #CB1C1C;
-  color: #fff;
-  width: 200px;
-  height: 40px;
-  border-radius: 10px;
-  margin-bottom: 5px;
-  transition: filter 0.3s ease;
-}
-
-.close:hover,
-.close:focus {
-  filter: brightness(1.2);
-}
-
-@media (max-width: 520px) {
-  .modal {
-    width: 400px;
-  }
-}
-
-@media (max-width: 410px) {
-  .modal {
-    width: 350px;
-  }
-}
-
-@media (max-width: 360px) {
-  .modal {
-    width: 300px;
-  }
-
-  .categoryName,
-  .taskDescription {
-    width: 220px;
-  }
-}
-
-@media (max-width: 320px) {
-  .modal {
-    width: 270px;
-  }
-}
