+++ src/components/Modals/Modals.module.css
@@ -1,0 +1,261 @@
+.modal {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  display: block;
+  color: #fff;
+  background-color: #a5cab6;
+  z-index: 10;
+  width: 500px;
+  height: 570px;
+  border-radius: 30px;
+  overflow: auto;
+}
+
+.modalcontent {
+  margin-top: 18px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+}
+
+input[type="date"] {
+  background-color: #e0e0e0;
+  padding: 10px;
+  padding-left: 8px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  font-size: 16px;
+  color: #333;
+  width: 130px;
+  border-radius: 5px;
+  transition: filter 0.3s ease;
+}
+
+input[type="date"]::-webkit-calendar-picker-indicator {
+  background-image: url('../../images/calendar.svg');
+  background-size: contain;
+  background-repeat: no-repeat;
+  opacity: 1;
+  color: transparent;
+  cursor: pointer;
+  width: 25px;
+  height: 25px;
+}
+
+input[type="date"]:focus {
+  border-color: #007bff;
+  outline: none;
+}
+
+input[type="date"]::-webkit-calendar-picker-indicator {
+  color: #007bff;
+  cursor: pointer;
+}
+
+input[type="date"]:hover {
+  filter: brightness(1.2);
+}
+
+.categoryName {
+  width: 100%;
+  height: 30px;
+  text-align: left;
+  padding-left: 0;
+  color: #000;
+  border: 0;
+  outline: none;
+  background-color: #a5cab6;
+  border-bottom: 1px solid #000;
+  width: 300px;
+}
+
+.categoryName::placeholder {
+  color: #000;
+  font-size: 16px;
+  font-weight: normal;
+  padding: 0;
+}
+
+.categoryName:focus {
+  border-color: #007bff;
+}
+
+.taskDescription {
+  background-color: #8fbc8f;
+  width: 300px;
+  color: #000;
+  height: 130px;
+  padding: 10px;
+  resize: none;
+  border-radius: 5px;
+  margin-bottom: 20px;
+}
+
+.taskDescription::placeholder {
+  color: #000;
+  font-size: 16px;
+  font-weight: normal;
+  padding-left: 5px;
+}
+
+.taskDescription:focus {
+  outline: none;
+  border: 1px solid #007bff;
+}
+
+.closeModalCategory {
+  background-color: #CB1C1C;
+  color: #fff;
+  width: 200px;
+  height: 40px;
+  border-radius: 10px;
+  margin-bottom: 10px;
+  transition: filter 0.3s ease;
+}
+
+.closeModalCategory:hover {
+  filter: brightness(1.2);
+}
+
+.close {
+  background-color: #CB1C1C;
+  color: #fff;
+  width: 200px;
+  height: 40px;
+  border-radius: 10px;
+  margin-bottom: 5px;
+  transition: filter 0.3s ease;
+}
+
+.close:hover,
+.close:focus {
+  filter: brightness(1.2);
+}
+
+input[type="date"] {
+  background-color: #e0e0e0;
+  padding: 10px;
+  padding-left: 8px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  font-size: 16px;
+  color: #333;
+  width: 130px;
+  border-radius: 5px;
+  transition: filter 0.3s ease;
+}
+
+input[type="date"]::-webkit-calendar-picker-indicator {
+  background-image: url('../../images/calendar.svg');
+  background-size: contain;
+  background-repeat: no-repeat;
+  opacity: 1;
+  color: transparent;
+  cursor: pointer;
+  width: 25px;
+  height: 25px;
+}
+
+input[type="date"]:focus {
+  border-color: #007bff;
+  outline: none;
+}
+
+input[type="date"]::-webkit-calendar-picker-indicator {
+  color: #007bff;
+  cursor: pointer;
+}
+
+input[type="date"]:hover {
+  filter: brightness(1.2);
+}
+
+.modalcontent input {
+  margin-bottom: 20px;
+  padding-left: 0px;
+}
+
+.modal select {
+  display: block;
+  margin-bottom: 20px;
+  border-radius: 10px;
+  background-color: #e0e0e0;
+  width: 200px;
+  height: 30px;
+}
+
+.modal select:focus {
+  outline: none;
+  border: 1px solid #007bff;
+}
+
+input[type="color"] {
+  width: 50px;
+  height: 25px;
+  cursor: pointer;
+}
+
+.closeModalCategory {
+  background-color: #CB1C1C;
+  color: #fff;
+  width: 200px;
+  height: 40px;
+  border-radius: 10px;
+  margin-bottom: 10px;
+  transition: filter 0.3s ease;
+}
+
+.closeModalCategory:hover {
+  filter: brightness(1.2);
+}
+
+option {
+  display: block;
+  text-align: center;
+  border-radius: 10px;
+  margin-bottom: 30px;
+}
+
+.modal_categoryAdd {
+  width: 400px;
+  height: 360px;
+}
+
+.description_color {
+  color: #000;
+  font-size: 16px;
+  margin-bottom: 20px;
+}
+
+@media (max-width: 520px) {
+  .modal {
+    width: 400px;
+  }
+}
+
+@media (max-width: 410px) {
+  .modal {
+    width: 350px;
+  }
+}
+
+@media (max-width: 360px) {
+  .modal {
+    width: 300px;
+  }
+
+  .categoryName,
+  .taskDescription {
+    width: 220px;
+  }
+}
+
+@media (max-width: 320px) {
+  .modal {
+    width: 270px;
+  }
+
+}
