--- src/components/Modals/TaskVariants/TaskVariants.module.css
@@ -1,42 +1,0 @@
-
-.modal {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  display: block;
-  color: #fff;
-  background-color: #a5cab6;
-  z-index: 10;
-  width: 500px;
-  height: 570px;
-  border-radius: 30px;
-  overflow: auto;
-}
-
-.modal_categoryAdd {
-  width: 400px;
-  height: 360px;
-}
-
-.modalcontent {
-  margin-top: 18px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-}
-
-.closeModalCategory {
-  background-color: #CB1C1C;
-  color: #fff;
-  width: 200px;
-  height: 40px;
-  border-radius: 10px;
-  margin-bottom: 10px;
-  transition: filter 0.3s ease;
-}
-
-.closeModalCategory:hover {
-  filter: brightness(1.2);
-}
